+++ MobileGarage/wearApp/src/debug/java/com/chriscartland/garage/wear/debug/ScreenshotStagesActivity.kt
@@ -68,6 +68,7 @@ class ScreenshotStagesActivity : ComponentActivity() {
                         onDoorTap = {},
                         onHoldStart = {},
                         onHoldEnd = {},
+                        onAnyTouch = {},
                         onSignInClick = {},
                     )
                 }